genturfa'i: allow cmene to be specified in nunjavni.
There was (and is) a routine called nunjavni-cmene.  That was the
only way to associate a name with a portion of the parse tree.  This
ability has been moved into each of the nunjavni, though
nunjavni-cmene still exists for the case of a non-terminal, which
needs to have the cmene associated at runtime instead of compile
time.

Remove all of the macro-expansion code for samselpla, making the
user specify the full lambda.  I did this both because I had a bug
in the macro I couldn't track down, and because the macro expansion
is not useful to the compiler, which needs to execute the parse tree
as/after it is constructed, and I did not want to run the macro
expander during the construction of a parser.

Update the tests to account for changes to the samselpla code.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21431 fca3e652-9b03-0410-8d7b-ac86a6ce46c4