genturfa'i: bootstrapping parser
The bootstrapping parser is not yet feature complete, it will not
run code embedded in a PEG parser.

It is otherwise working, and I've converted each non-coding
containing test to use the PEG parser as well as the original
hand-written grammar.

Add a program, genturfahi-peg, which parses PEG grammar files and
outputs a .scm file containing a parser.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21478 fca3e652-9b03-0410-8d7b-ac86a6ce46c4