genturfa'i: remove genturfahi-bootstrap parameter.
This parameter was used to interface the output of the parser with
the rest of the code.  I've rewritten the compiler to not use dynamic
variables, and instead generate a code stub that can be used
directly.

I was running into problems with the previous code, when I would
compile my PEG compiler, then compile a PEG grammar, and then want
to compile another PEG grammar.  The PEG compiler was being
overwritten with the compiler for the grammar.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21571 fca3e652-9b03-0410-8d7b-ac86a6ce46c4