genturfa'i: update bootstrap compiler to (better) handle regular expressions.
This isn't the final syntax, but it is enough to get the .peg
version of the re test working.

Note that I am reading and writing the compiler output to a port
before evaluating it.  This is to avoid a bug or hygentic macro
issue, I'm not sure which yet.  The straight PEG version of these
tests are not yet working and are therefor commented out.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21573 fca3e652-9b03-0410-8d7b-ac86a6ce46c4