genturfa'i: move samselpla tags inside jonai blocks.
I have reworked the handling of code and #!key arguments.
non-terminals no longer automatically get a #!key.  code is
associated with unary-rules, allowing multiple code blocks
per non-terminal expression if that expression has an ordered-choice
operator.

I've begun to clean up the code to make it more compatible with the
eventual compiler output as well, transforming inline lambda
expressions to a direct call to the code in samselpla.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21585 fca3e652-9b03-0410-8d7b-ac86a6ce46c4