genturfa'i: Add mex.peg parsing test.
This is the final test that I had not converted to peg.  The
bootstrap compiler is now compiling PEG grammars for each input
test.  The next step is to have it compile itself.

I'm experiencing the same problem here as I am in the re.peg test,
which I've filed as ticket 433:

  https://bugs.call-cc.org/ticket/433

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21586 fca3e652-9b03-0410-8d7b-ac86a6ce46c4