genturfa'i: move string->keyword into nunjavni-samselpla
I had focused on generating proper keyword symbols in the compiler.
As these symbols vary by scheme implementation, and because a symbol
is not the same as a keyword, convert all of this code to use
strings instead, which are converted to keywords as late as
possible: during runtime after a successful parse.

I could in the future do this when I'm generating rules, which would
make the parser faster at the expense of exporting the code to
convert strings to keywords all over the nunjavni.scm file.

Passing strings rather than almost-but-no-quite properly formed
symbols around the compiler should help me avoid future problems
related to keyword handling, and make the code easier to port to
other Scheme systems.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21604 fca3e652-9b03-0410-8d7b-ac86a6ce46c4