genturfa'i: checkpoint secuxna handling
This code still doesn't work, but this is my current checkpoint.

The problem is that I need these options to run before I generate
the rest of the parse tree, but the code is not executed until
after the parse tree is generated.

I could wrap all of the parsing code up in a generator, and call
the generator code after I've executed these options.  That
significantly complicates the PEG parsing code, so I'm going to
wait for further inspiration.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21638 fca3e652-9b03-0410-8d7b-ac86a6ce46c4