genturfa'i: recognize '()' (empty grouping operator) as the empty string.
Previously the compiler did not emit any production to match the
empty string.  The test case worked around this by literally
specifying the empty string, but that can now be customized by the
user.

Further, the Lojban grammar has an occurance of '()', there is no
reason it isn't valid PEG grammar.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21786 fca3e652-9b03-0410-8d7b-ac86a6ce46c4