genturfa'i: move PEG parser to its own library.
Since the PEG parser has to be initialized at runtime, there is a
cost to including it in genturfa'i.  I want genturfa'i to load as
fast as possible, so I'm moving all of the PEG code, which isn't
required at runtime like the rest of genturfa'i, to the
genturfahi-peg library.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21822 fca3e652-9b03-0410-8d7b-ac86a6ce46c4