genturfa'i: add , and ' operators to grammar.
This is a major cleanup of the interface between the parser and
code, allowing a mixture of #!key and positional arguments, letting
the user specify which productions in the grammar modify the parse
tree and which don't.

By default, literal strings and characters no longer add elements to
the parse tree.  As well, the predicate operations no longer modify
the parse tree either.

The , operator can be used in front of a string or character so that
it is included in the parse tree.  Conversely, the ' operator can be
used to mark a rule so it does not modify the parse tree.

These together allow a much cleaner and more flexible interface
between the grammar and the code that operates on the parse tree.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21888 fca3e652-9b03-0410-8d7b-ac86a6ce46c4