genturfa'i: add -| operator for immediate code execution.
This is used to execute the header code in genturfahi.peg before
parsing the rest of the grammar.

Because of the rules of scheme and the implementation of of
genturfa'i, I can't guarantee that the parse tree is executed in any
particular order, save that the leaves will be executed before their
branches.

Even if I change genturfa'i to guarantee a certain kind of order, I
can't change the Scheme standard, which does not requires parameters
to a function to be evaluated in any particular order.

This change executes the code during the parse.  It may execute
multiple times, though with memoization turned on this should not be
true.  Regardless, I don't really specify the number of times this
code is executed, and have limited my operations to those that can
be executed multiple times with no side-effects.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21950 fca3e652-9b03-0410-8d7b-ac86a6ce46c4