genturfa'i: turn on all optimization settings.
I've uncommented all of the (declare ...) options, as I think the
code is ready to run with these optimizations.

I've been using this set when I run genturfa'i for jbogenturfa'i,
and the errors I've seen all turn out to be user-error.

I do, with these settings, permit malformed input to the PEG parser
to result in error messages (or core dumps) from the compiler.  I'm
happy, with the scale genturfa'i is at now, to leave this as it is.

A more user-friendly program would catch these errors and report
something domain-specific.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@21952 fca3e652-9b03-0410-8d7b-ac86a6ce46c4