genturfa'i: add debug and profile support to parser.
This is a fairly substantial patch, adding all of the framework to
produce debug output and profiling information from the parser.
When these options are not enabled, there is a small initialization
penalty and no runtime penalty to the parser.

When debug in enabled, a file is written containing a symbolic
expression describing the path taken by the parser while it matches
the input.

When profile is enabled, timing for non-terminal rules and operators
is written as an association list.  This information isn't perfect
yet, as it doesn't subtract containing rules from the timing of
outer rules.  For purposes of debugging performance, however, it is
good enough.  It allowed me to find some performance issues in the
memoization code, a patch for which is coming up.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@22102 fca3e652-9b03-0410-8d7b-ac86a6ce46c4