genturfa'i: !. operator, for consistency with !, does not modify parse.
I had been returning the sentinel character from this rule, but as
the rule is a not predicate, it is not appropriate for it to be
returning anything at all.

By default I no longer modify the tree, and there is no way in the
compiler for the user to access the feature turning it back on.
(Like all ! productions.)

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@22350 fca3e652-9b03-0410-8d7b-ac86a6ce46c4