genturfa'i: allow -| operator to control parse.
This routine executes code immediately.  I now permit the executed
code to return a value which is passed to the rest of the parse.  If
the value matches |secuxna-nonmatch-token|, the non-match path is
taken instead.

I'm going to use this in jbogenturfa'i for processing ZOI, which has
the extra-gramatical requirement that two tokens both match a rule
and that they match each other.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@22378 fca3e652-9b03-0410-8d7b-ac86a6ce46c4