genturfa'i: keep -> and ` from fighting over argument format.
the sequence operator will sometimes return a single result, if all
but one argument are marked as not modifying the parse tree.

the -> operator needs arguments in list form for use by apply.

The heuristic I was using to satisfy both of these requirements
failed when the single result from the sequence operation was itself
a list.

Create a new parse tree method that does not convert single-element
lists to atoms for use by the -> operator, obviating the need to
convert atoms back to single element lists, which was ambiguous.

Check in a test that handles each variant of this use.  Fix a bug
in the grammar compiler which was working around this problem.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@22420 fca3e652-9b03-0410-8d7b-ac86a6ce46c4