genturfa'i: refactor behavior of grouping "()" operator.
Structurally, this parser generator has the following components:

  javni:    the individual elements that make up a production, like
            '*', '?', &c.
  valsi:    the tokens that are constructed when a rule matches and
            placed into the parse tree.
  compiler: the translation from PEG to scm, which manages the
            interface between all of the javni and they way their
            created valsi are constructed into a parse tree.

This commit formalizes the interation between rule creation and
token creation.  This was done so that the grouping operator doesn't
affect the resulting parse tree.  The following rules (and those
like them) are the same:

  a <- a b c
  a <- (a b) c
  a <- (a (b c)

When a grouping operator is encountered, the result rule is spliced
into the sequence it surrounds.  Formalizing this behavior required
handling a lot of edge cases, so this commit is fairly large.

Noteworthy changes:

 * group splicing is handled by the porjahe and porsumti flags,
   which describe whether to generate a list to splice or whether
   to accept a list to splice from a contained rule.
 * I'm now using the matchable egg when generating valsi rules.
   These rules are generated based on a combination of several
   flags, and the matchable egg allows the structure of these
   choices to be seen.
 * the je production in the PEG grammar acts as a glue interface,
   generating code based on a fairly broad look at incoming rules.
   I'm also using the matchable egg here.

This code is smaller (I think!) and not as complex.  In particular,
there are fewer choice points in most of the code, the flags
generating the correct behavior which is then assumed by the rest of
the system, with almost all of the glue layer now in the compiler.

I like the interface between these systems much better now, and I
believe the runtime performance of the parser is improved by
removing the checks I had added for various states.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@22502 fca3e652-9b03-0410-8d7b-ac86a6ce46c4