genturfa'i: teach ()? to handle grouped rules that don't modify the tree.
The optional rule, when used to match a group, returns as many
empty-strings as there are elements in the group.

If any of the elements in the group returned the token indicating
they shouldn't modify the tree, ()? would not honor it, returning
the empty string instead.

Detect cases of non-parse tree modifying rules and only return as
many empty string tokens as we have parse tree modifying rules.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@22580 fca3e652-9b03-0410-8d7b-ac86a6ce46c4