genturfa'i: allow repetition of ordered choice to include ` rules.
This production:

  a <- (,#\a / `,#\b)*

Has one choice which modifies the tree, and one choice that doesn't.
A bug in genturfa'i was exposing the special token that indicates
the tree should not be modified, effectively making the above rule
useless.

I've fixed the bug here and briefly checked for other instances of
it.  I believe this is the only one.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@22789 fca3e652-9b03-0410-8d7b-ac86a6ce46c4