genturfahi: in optional ordered choice, take max count rather than sum
Since only one match of an ordered choice can be returned, we should
not sum all of the possible elements that could be returned.
Instead, take the maximum of any of the individual rules.

As well, if the choice only has a single argument, don't both: it's
a single element and we'll treat it as such.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@25411 fca3e652-9b03-0410-8d7b-ac86a6ce46c4