genturfahi: fix grouping operator when grouped around a single element.
The grammar:

a <- (,#\a #\b)?

Was returning "" when there was no match, and returning '(#\a) when
there was a match.  The expectation is that it return #\a instead,
as there is only a single parse-tree modifying element in the group.

Correct this behavior so that the single token is returned, rather
than a list of that single token.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@25412 fca3e652-9b03-0410-8d7b-ac86a6ce46c4