genturfa'i: use list, rather than vector, for parser input.
This converts several of the now o(n) routines back to o(1).  I
noticed the performance hit, just testing as I converted this
code.

Rather than storing a vector and an index, we store the current
position in the list and advance it as we parse.  The grouping
code is now immeasurably slower.  You won't notice.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/genturfahi/trunk@25738 fca3e652-9b03-0410-8d7b-ac86a6ce46c4