add utf8 support to PEG grammar.
Much of this support has been in bootstrap.scm for some time.  I
accidentally deleted genturfahi.peg the file that generates boostrap.scm.

This is both a partial restoration of the features from
bootstrap.scm in to genturfahi.peg, but also formally publishing and
making them available in the egg.