add sentinel key argument to genturfahi and genturfahi*
sentinel, like the non-match token, is a parse-time, not
compile-time variable.  This is one area that is pretty broken
right now, with a lot of confusion in the code about when a
parameter is set or not.

The stack-related changes that recently went into the memoization
code is also related to this.

At any rate, fix what I can: allow the sentinel to be set at parse
time so we can parse grammars that have the sentinel character in
them.

It isn't clear to me the sentinel character is as important as it
used to be.  The idea of a character as a sentinel comes from before
utf8 support, when I parsed the string buffer rather than splitting
it out into a list.  I may be able to use a gensym as a sentinel,
which would allow me to completely remove the user-interface portion
of sentinel handling by never having the possibility of the sentinel
appearing in the input.

I have investigation to do here still, so save that for a later
commit.