Add support for parsing nul characters to grammar.
character classes now support [\0] to represent the nul string.  The
character #\nul is now parsed as a nul character, and "\0" is an
embedded nul in a string.

extend the string-escape test to handle this case, including testing
support for non-nul sentinels.

I'm leaving off support/testing for a grammar that also does EOF
parsing, as that bumps up against bugs in how and when parameters
are evaluated.